// start-microfrontends/apps/se/src/app/application/tramites/80104/models/nuevo-programa-industrial.model.ts
export interface ProveedorClienteDatosTabla {
  idProveedor: number;
  paisOrigen: string;
  rfcProveedor: string;
  razonProveedor: string;
  paisDestino: string;
  rfcClinte: string;
  razonSocial: string;
  domicilio: string;
  testado: boolean;
  idProductoP: number;
  descTestado: string;
}
// ...
export interface DatosParaNavegar {
   encabezadoAnexoII?: string;
   encabezadoTipo?: string;
   encabezadoCategoria?: string;
   encabezadoDescripcionComercial?: string;
   encabezadoVolumenMensual?: string;
   encabezadoVolumenAnual?: string;
   encabezadoValorEnMonedaMensual?: string;
   encabezadoValorEnMonedaAnual?: string;
}